+++ PacletInfo.m
@@ -1,10 +1,10 @@
 (* Paclet Info File *)
 
-(* created 2019/03/16*)
+(* created 2019/03/20*)
 
 Paclet[
     Name -> "MaXrd",
-    Version -> "1.2.0",
+    Version -> "1.3.0",
     MathematicaVersion -> "11.3+",
     Description -> "Database and utilities related to crystallography and X-ray scattering.",
     Creator -> "Stian Penev Ramsnes",
@@ -13,7 +13,9 @@ Paclet[
         {
             {"Application", Context -> "MaXrd`"}, 
             {"Documentation", Language -> "English", MainPage -> "Guides/MaXrd"}, 
-            {"Documentation", Resources -> {"Guides/MaXrd"} }
+            {"Documentation", Resources -> 
+                {"Guides/MaXrd"}
+            }
         }
 ]